Created a working preferences page. Sort of. Right now the colors are working. Has an amazing animation when you change between the colors, and also can remember what color you choose between app launches, meaning that the storing is working.
Bugs:
No profile picture.
No buttons for the account actions
+ Any bug listed in a previous commit.

To Do: Add the profile button to the work page in order to get to the settings page. Make sure that the color update when leaving the settings page.
This could be guaranteed by just creating a new modal of the list view, and displaying the new one, trashing the old one, which would save the memory in the preferences view, and would eliminate having to think about refreshing