The new item is working on the work page. Shows them. Does not hide them, and they do not do anything
Bugs:
Pop-ups don't go away.
Pop-ups dont work

Suggestions:
Should add all of the pop-up options and the width to the initializer, that way you dont forget to include them.

ToDo:
Get the pop-ups to go away. Get them to call a function. This function doesn't have to be the pop-up one.