Was able to create the tab bar with showing and hiding the labels of the tab bar. When you touch the button, it hides it,
Bugs:
Cant hide them by touching anywhere outside of the view. Should be easy enough
The height of the extension is hardcoded. Should be able to just take the botton corner from the frame, and go 12 further then that.

ToDo:
Get it to hide when continuing a scroll on the tableview and also when the buttons aren't hit, but you hit the green area.