Was able to create the pop-up series to show the new class, and new assignment. The new assignment creates an assignment, and saves it to local storage. Also, the work view fetches the work that you have
Should go into LocalActions.WorkObjects.All.fetch() and make it not return the object if it is empty, cause now we have the bottom cell anyway.

Animations SUCK, need to make an improvement in that department. Possibly the conformation, and gesture controls in the making of new objects.

ToDo:
Delete locally stored objects