The work images are working. They are saved to storage, references both in firebase and locally on the work files. When the works are downloaded when a user signs in, the images associated with them are also signed in. On top of that, the works are saving themselves, the classes are also working. The App will not crash right now. When adding a new image, and this might just be when on extremely slow wifi, it takes forever for the completion handler to run. The assignments can also be tapped, as they register Italy nd will move in. Everythignis taken care of when you deliberately sign out.
ToDo:
Fix the animation of the new item popup, should be as simple as starting with alpha of 0 and fading in once the veiw is there, and the thing is in the middle of popping up
Start a loading screen
Put the notes view in, do not make functional yet.
In the picture collection view, make it possible to have a card which asks if you want a new one, will have the same methods as the tab bar functions.