Was able to make it so you can now organize the list. Also, added a status bar, to say things such as when the last pull down was, and what is going on in the app. Could also be used to show if someone is typing to you later when we implement the messaging feature.
To Do:
Make it so that you can draw in images. I think this is going to be a good touch. Do something where the picture saves, and you are able to open it in another view controller. This one has the option to just view it, but also you can draw on it. This image will then ask you if you want to save it when you are done. if you say yes, then it updates the image with the new drawing overlaying it!

BTW, this app is fracking amazing. It is going to be life changing