Was able to make deletions work very well. Implemented the class deletion and work deletion. Image deletion for work might not be needed, will do at a later date. To manage classes, go to preferences page, and click "Manage classes". This view is ok for now, but it is definitely going to need to be changed.
Noted for future to do:
Need to make the notification bar ask if you are sure to delete the work, and then make sure to change the icon to an undo button, and change the action during that time

ToDo:
Work on creating working icons. The app icon is not included in this, but the images that are found throughout the app and the images at the bottom of the tableviews are a must.